Setup frontend communication
