+++ frontend/question-service/src/Question.tsx
@@ -1,0 +1,9 @@
+interface Question {
+  qId: number;
+  qTitle: string;
+  qDescription: string;
+  qCategory: string[];
+  qComplexity: string;
+}
+
+export default Question;
