--- frontend/question-service/src/axios.js
@@ -1,6 +1,0 @@
-import axios from "axios";
-const instance = axios.create({
-  baseURL: "http://localhost:9000",
-});
-
-export default instance;
