+++ frontend/question-service/src/axios.tsx
@@ -1,0 +1,7 @@
+import axios, { AxiosInstance } from "axios";
+
+const instance: AxiosInstance = axios.create({
+  baseURL: "http://localhost:9000",
+});
+
+export default instance;
